# rewrite performance 'tests' to use criterium